+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "flex-headers",
-  "version": "1.7.3",
+  "version": "1.7.4",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^6.9.1",
