+++ src/constants.ts
@@ -2,7 +2,8 @@ export const SETTINGS_V3_META_KEY = "settings_v3_meta";
 export const PAGE_KEY_PREFIX = "page_";
 export const SELECTED_PAGE_KEY = "selectedPage";
 export const SETTINGS_SAVE_DEBOUNCE_TIME = 1000;
-export const SYNC_INTERVAL = 10000; // Sync from local to sync storage every 30 seconds
+export const SYNC_INTERVAL = 10000; // Sync from local to sync storage every 10 seconds
+export const SYNC_ITEM_BYTE_LIMIT = 8192; // chrome.storage.sync's per-item quota (QUOTA_BYTES_PER_ITEM)
 export const LAST_SYNC_TIME_KEY = "last_sync_time"; // Track when we last pushed local data to sync storage
 export const LAST_MERGE_TIME_KEY = "last_merge_time"; // Set only when the background script merges remote pages into local storage
 export const LOCAL_MODIFIED_TIME_KEY = "localModifiedTime"; // Bumped on every local write - compared against LAST_SYNC_TIME_KEY to tell the user "sync pending" vs "synced"
